+++ driver-core/src/main/java/com/datastax/driver/core/TokenRange.java
@@ -213,6 +213,16 @@ public final class TokenRange implements Comparable<TokenRange> {
         return intersected;
     }
 
+    /**
+     * Checks whether this range contains a given token.
+     *
+     * @param token the token to check for.
+     * @return whether this range contains the token, i.e. {@code range.start &lt; token &lt;= range.end}.
+     */
+    public boolean contains(Token token) {
+        return contains(token, false);
+    }
+
     // isStart handles the case where the token is the start of another range, for example:
     // * ]1,2] contains 2, but it does not contain the start of ]2,3]
     // * ]1,2] does not contain 1, but it contains the start of ]1,3]